pangb 1.2.1
- Updated LICENSE
- New repository structure
- Start conversion to fixed-lenght types
- A little code cleanup
- Introduced basic support to mbc3

Please note that mbc3 support isn't complete and it's disabled by default

+++ src/gb-emu/gb-opcodes.c
@@ -2,7 +2,7 @@
 #include "gb.h"
 #include "gb-opcode.h"
 
-unsigned int executeOpcode(gb *cpu, BYTE opcode){
+BYTE executeOpcode(gb *cpu, BYTE opcode){
     BYTE val = 0;
     SIGNED_BYTE s_val = 0;
     //printf("PC %x = %x\n",cpu->progCounter-1, opcode);
